refactor: 검색바 컴포넌트화 및 UI 변경
