+++ src/pages/Store/StorePage/components/SearchBar/SearchBar.module.scss
@@ -1,0 +1,72 @@
+@use "src/utils/scss/media";
+
+.search_bar {
+  text-align: center;
+  position: relative;
+
+  @include media.media-breakpoint(mobile) {
+    justify-content: center;
+    align-items: center;
+    display: flex;
+  }
+
+  &__input {
+    font-family: NanumBarunGothic, serif;
+    border: solid 1px #d2dae2;
+    height: 45px;
+    width: 1068px;
+    font-size: 15px;
+    line-height: 1.6;
+    letter-spacing: -0.09px;
+    padding-left: 20px;
+    padding-right: 40px;
+
+    @include media.media-breakpoint(mobile) {
+      width: calc(100% - 95px);
+      height: 40px;
+      font-size: 14px;
+      border: none;
+      background-color: #f5f5f5;
+      line-height: 1.6;
+      letter-spacing: -0.09px;
+      padding-left: 16px;
+      padding-right: 40px;
+    }
+
+    &::placeholder {
+      color: #d2dae2;
+
+      @include media.media-breakpoint(mobile) {
+        color: #727272;
+      }
+    }
+  }
+
+  &__icon {
+    position: absolute;
+    height: 24px;
+    right: 11px;
+    bottom: 13px;
+    z-index: 1;
+    border: 0;
+    background-color: #fff;
+    cursor: pointer;
+
+    @include media.media-breakpoint(mobile) {
+      right: 33px;
+      bottom: 8px;
+    }
+  }
+}
+
+.search-icon {
+  height: 24px;
+  z-index: 0;
+  border: 0;
+  background-color: #fff;
+
+  @include media.media-breakpoint(mobile) {
+    background-color: #f5f5f5;
+    color: red;
+  }
+}
